patrol: parseInt(raw) with bitrate values like '1.5M' returns 1 instead of 1.5, causing incorrect scaling. parseInt stops at first non-digit character and doesn't handle decimal points. (#90)

+++ js/vidconv-engine.js
@@ -113,7 +113,7 @@ function applyQualityToArgs(args, quality) {
   if (bvIdx >= 0) {
     const raw = out[bvIdx + 1];
     const isM = raw.includes("M");
-    const baseK = parseInt(raw) * (isM ? 1000 : 1);
+    const baseK = parseFloat(raw) * (isM ? 1000 : 1);
     const scale = quality === "medium" ? 0.6 : 0.35;
     const scaledK = Math.round(baseK * scale);
     out[bvIdx + 1] =
